+++ tests/test_write.py
@@ -4,6 +4,7 @@ from typing import Any
 import pyarrow.parquet
 import rustac
 import stac_geoparquet
+from rustac.store import LocalStore
 
 
 async def test_write(item: dict[str, Any], tmp_path: Path) -> None:
@@ -19,3 +20,10 @@ async def test_write_compressed(item: dict[str, Any], tmp_path: Path) -> None:
     await rustac.write(path, [item])
     metadata = pyarrow.parquet.read_metadata(path)
     assert metadata.row_group(0).column(0).compression == "SNAPPY"
+
+
+async def test_write_store(item: dict[str, Any], tmp_path: Path) -> None:
+    store = LocalStore(prefix=tmp_path)
+    await rustac.write("item.json", item, store=store)
+    read_item = await rustac.read("item.json", store=store)
+    assert item["id"] == read_item["id"]